Add abstract DatatokenBase.
