+++ ocean_lib/models/data_nft.py
@@ -12,7 +12,7 @@ from brownie import network
 from enforce_typing import enforce_types
 from web3 import Web3
 
-from ocean_lib.models.datatoken import Datatoken, DatatokenArguments
+from ocean_lib.models.datatoken_base import DatatokenArguments, DatatokenBase
 from ocean_lib.ocean.util import (
     create_checksum,
     get_address_of_type,
@@ -22,7 +22,6 @@ from ocean_lib.ocean.util import (
 from ocean_lib.web3_internal.constants import ZERO_ADDRESS
 from ocean_lib.web3_internal.contract_base import ContractBase
 from ocean_lib.web3_internal.utils import check_network
-
 
 """
 def addManager(address: str) -> None:
@@ -290,7 +289,7 @@ class Flags(IntFlag):
 class DataNFT(ContractBase):
     CONTRACT_NAME = "ERC721Template"
 
-    def create_datatoken(self, tx_dict, *args, **kwargs) -> Datatoken:
+    def create_datatoken(self, tx_dict, *args, **kwargs) -> DatatokenBase:
         datatoken_args = get_args_object(args, kwargs, DatatokenArguments)
 
         return datatoken_args.create_datatoken(self, tx_dict)
