+++ ocean_lib/models/test/test_datatoken.py
@@ -6,7 +6,7 @@ import pytest
 from brownie import network
 from web3.main import Web3
 
-from ocean_lib.models.datatoken import DatatokenRoles, TokenFeeInfo
+from ocean_lib.models.datatoken_base import DatatokenRoles, TokenFeeInfo
 from ocean_lib.ocean.util import get_address_of_type, to_wei
 from ocean_lib.web3_internal.constants import MAX_UINT256
 from tests.resources.helper_functions import get_mock_provider_fees
