+++ ocean_lib/models/test/test_datatoken_order_both_templates.py
@@ -6,7 +6,7 @@ from datetime import datetime
 
 import pytest
 
-from ocean_lib.models.datatoken import Datatoken, TokenFeeInfo
+from ocean_lib.models.datatoken_base import DatatokenBase, TokenFeeInfo
 from ocean_lib.ocean.util import from_wei, get_address_of_type, to_wei
 from ocean_lib.web3_internal.constants import MAX_UINT256
 from tests.resources.helper_functions import deploy_erc721_erc20, get_mock_provider_fees
@@ -24,8 +24,8 @@ def test_dispense_and_order_with_non_defaults(
     """Tests dispense_and_order function of the Datatoken Enterprise"""
     _, DT = deploy_erc721_erc20(config, publisher_wallet, publisher_wallet, 2)
 
-    USDC = Datatoken(config, get_address_of_type(config, "MockUSDC"))
-    DAI = Datatoken(config, get_address_of_type(config, "MockDAI"))
+    USDC = DatatokenBase.get_typed(config, get_address_of_type(config, "MockUSDC"))
+    DAI = DatatokenBase.get_typed(config, get_address_of_type(config, "MockDAI"))
 
     _ = DT.create_dispenser(
         max_tokens=to_wei(1), max_balance=to_wei(1), tx_dict={"from": publisher_wallet}
@@ -157,8 +157,8 @@ def test_buy_DT_and_order(
         config, publisher_wallet, publisher_wallet, template_index
     )
 
-    USDC = Datatoken(config, get_address_of_type(config, "MockUSDC"))
-    DAI = Datatoken(config, get_address_of_type(config, "MockDAI"))
+    USDC = DatatokenBase.get_typed(config, get_address_of_type(config, "MockUSDC"))
+    DAI = DatatokenBase.get_typed(config, get_address_of_type(config, "MockDAI"))
 
     exchange = DT.create_exchange(
         rate=to_wei(1),
