+++ ocean_lib/ocean/mint_fake_ocean.py
@@ -7,7 +7,7 @@ import os
 from brownie.network import accounts
 from enforce_typing import enforce_types
 
-from ocean_lib.models.datatoken import Datatoken
+from ocean_lib.models.datatoken_base import DatatokenBase
 from ocean_lib.ocean.util import get_ocean_token_address, to_wei
 
 
@@ -20,7 +20,9 @@ def mint_fake_OCEAN(config: dict) -> None:
     """
     deployer_wallet = accounts.add(os.environ.get("FACTORY_DEPLOYER_PRIVATE_KEY"))
 
-    OCEAN_token = Datatoken(config, address=get_ocean_token_address(config))
+    OCEAN_token = DatatokenBase.get_typed(
+        config, address=get_ocean_token_address(config)
+    )
     amt_distribute = to_wei(2000)
     OCEAN_token.mint(deployer_wallet.address, to_wei(20000), {"from": deployer_wallet})
     for key_label in ["TEST_PRIVATE_KEY1", "TEST_PRIVATE_KEY2", "TEST_PRIVATE_KEY3"]:
