+++ tests/integration/ganache/test_onchain.py
@@ -10,7 +10,7 @@ from brownie import network
 
 from ocean_lib.agreements.service_types import ServiceTypes
 from ocean_lib.data_provider.data_service_provider import DataServiceProvider
-from ocean_lib.models.datatoken import Datatoken, DatatokenArguments
+from ocean_lib.models.datatoken_base import DatatokenArguments, DatatokenBase
 from ocean_lib.ocean.ocean_assets import OceanAssets
 from ocean_lib.ocean.util import get_address_of_type, to_wei
 from ocean_lib.structures.file_objects import SmartContractCall
@@ -77,7 +77,7 @@ def test_consume_parametrized_onchain_data(
     assert ddo.datatokens[0]["symbol"] == "DT1"
 
     service = get_first_service_by_type(ddo, ServiceTypes.ASSET_ACCESS)
-    dt = Datatoken(config, ddo.datatokens[0]["address"])
+    dt = DatatokenBase.get_typed(config, ddo.datatokens[0]["address"])
 
     # Mint 50 datatokens in consumer wallet from publisher. Max cap = 100
     dt.mint(
