+++ tests/resources/ddo_helpers.py
@@ -12,7 +12,7 @@ import requests
 from ocean_lib.agreements.service_types import ServiceTypes
 from ocean_lib.assets.ddo import DDO
 from ocean_lib.data_provider.data_service_provider import DataServiceProvider
-from ocean_lib.models.datatoken import DatatokenArguments
+from ocean_lib.models.datatoken_base import DatatokenArguments
 from ocean_lib.ocean.ocean import Ocean
 from ocean_lib.services.service import Service
 from ocean_lib.structures.algorithm_metadata import AlgorithmMetadata
