+++ tests/resources/helper_functions.py
@@ -21,7 +21,7 @@ from web3 import Web3
 from ocean_lib.example_config import get_config_dict
 from ocean_lib.models.data_nft import DataNFT
 from ocean_lib.models.data_nft_factory import DataNFTFactoryContract
-from ocean_lib.models.datatoken import Datatoken
+from ocean_lib.models.datatoken_base import DatatokenBase
 from ocean_lib.ocean.ocean import Ocean
 from ocean_lib.ocean.util import get_address_of_type, to_wei
 from ocean_lib.structures.file_objects import FilesTypeFactory
@@ -153,11 +153,11 @@ def deploy_erc721_erc20(
     data_nft_publisher,
     datatoken_minter: Optional = None,
     template_index: Optional[int] = 1,
-) -> Union[DataNFT, Tuple[DataNFT, Datatoken]]:
+) -> Union[DataNFT, Tuple[DataNFT, DatatokenBase]]:
     """Helper function to deploy an DataNFT using data_nft_publisher Wallet
     and an Datatoken data token with the newly DataNFT using datatoken_minter Wallet
     if the wallet is provided.
-    :rtype: Union[DataNFT, Tuple[DataNFT, Datatoken]]
+    :rtype: Union[DataNFT, Tuple[DataNFT, DatatokenBase]]
     """
 
     data_nft_factory = DataNFTFactoryContract(
@@ -204,7 +204,7 @@ def send_mock_usdc_to_address(config: dict, recipient: str, amount: int) -> int:
     """
     factory_deployer = get_factory_deployer_wallet(config)
 
-    mock_usdc = Datatoken(config, get_address_of_type(config, "MockUSDC"))
+    mock_usdc = DatatokenBase.get_typed(config, get_address_of_type(config, "MockUSDC"))
     initial_recipient_balance = mock_usdc.balanceOf(recipient)
 
     if mock_usdc.balanceOf(factory_deployer) >= amount:
@@ -226,7 +226,7 @@ def transfer_bt_if_balance_lte(
     is less or equal to min_balance and from_wallet has enough ocean balance to send.
     Returns the transferred ocean amount.
     """
-    base_token = Datatoken(config, bt_address)
+    base_token = DatatokenBase.get_typed(config, bt_address)
     initial_recipient_balance = base_token.balanceOf(recipient)
     if (
         initial_recipient_balance <= min_balance
